Validate shaders on travis

+++ source/Maze.cpp
@@ -28,16 +28,16 @@ Maze::Maze(const size_t& size)
 std::vector<WallSlot> Maze::walls()
 {
     std::vector<WallSlot> result;
-    for (int y = 0; y < m_wallsX.size(); ++y)
+    for (size_t y = 0; y < m_wallsX.size(); ++y)
     {
-        for (int x = 0; x < m_wallsX[y].size(); ++x)
+        for (size_t x = 0; x < m_wallsX[y].size(); ++x)
         {
             result.push_back({{x, y}, {x + 1, y}, m_wallsX[y][x]});
         }
     }
-    for (int x = 0; x < m_wallsY.size(); ++x)
+    for (size_t x = 0; x < m_wallsY.size(); ++x)
     {
-        for (int y = 0; y < m_wallsX[x].size(); ++y)
+        for (size_t y = 0; y < m_wallsX[x].size(); ++y)
         {
             result.push_back({{x, y}, {x, y + 1}, m_wallsY[x][y]});
         }
